Create LineItUp.c
This tool will help the programmer to convert the content of one file into another file in one line , and also the length of the content.We just have to provide the name of the file whose content we want to convert , and run the programme. The One Liner text will be Saved in file named as tmp.tmp.